+++ pysnmp/hlapi/asyncio/lcd.py
@@ -275,11 +275,13 @@ class NotificationOriginatorLcdConfigurator(AbstractLcdConfigurator):
             cache["auth"][authDataKey] = authDataX, subTree, useCount + 1
         else:
             subTree = (1, 3, 6)
-            config.addTrapUser(
+            config.addVacmUser(
                 snmpEngine,
                 authData.securityModel,
                 authData.securityName,
                 authData.securityLevel,
+                (),
+                (),
                 subTree,
                 contextName=contextName,
             )