+++ pysnmp/proto/errind.py
@@ -238,9 +238,6 @@ class UnsupportedSecurityLevel(ErrorIndication):
     pass
 
 
-# backward compatibility plug
-UnsupportedSecLevel = UnsupportedSecurityLevel
-
 unsupportedSecurityLevel = UnsupportedSecurityLevel("Unsupported SNMP security level")