+++ pysnmp/proto/secmod/rfc2576.py
@@ -8,7 +8,7 @@ import sys
 from pyasn1.codec.ber import encoder
 from pyasn1.error import PyAsn1Error
 from pysnmp.proto.secmod import base
-from pysnmp.carrier.asyncio.dgram import udp, udp6, unix
+from pysnmp.carrier.asyncio.dgram import udp, udp6
 from pysnmp.smi.error import NoSuchInstanceError
 from pysnmp.proto import errind, error
 from pysnmp import debug
@@ -207,12 +207,6 @@ class SnmpV1SecurityModel(base.AbstractSecurityModel):
                         "TRANSPORT-ADDRESS-MIB", "TransportAddressIPv6"
                     )
                     targetAddrTAddress = tuple(TransportAddressIPv6(targetAddrTAddress))
-                elif (
-                    targetAddrTDomain[: len(unix.SNMP_LOCAL_DOMAIN)]
-                    == unix.SNMP_LOCAL_DOMAIN
-                ):
-                    targetAddrTAddress = str(targetAddrTAddress)
-
                 targetAddr = targetAddrTDomain, targetAddrTAddress
                 targetAddrTagList = snmpTargetAddrTagList.getNode(
                     snmpTargetAddrTagList.name + instId