+++ CHANGES.rst
@@ -2,6 +2,9 @@ Revision 7.0.0, released on Aug 22, 2024
 -----------------------------------------
 
 - Applied PEP8 recommended names. This breaks backward compatibility.
+- Many really old backward-compatibility code snippets removed.
+  Most importantly, everything related to (non-standard) UNIX domain socket
+  transport are gone.
 
 Revision 6.2.5, released on Aug 10, 2024
 -----------------------------------------